Correct solution for Exercise 6 from Chapter 7

+++ ch07/exercises/c07e06/c07e06.c
@@ -5,12 +5,12 @@ int main(void)
     char days_in_month = 31;
     short days_in_year = 366;
     short minutes_in_day = 60 * 24;
-    int seconds_in_day = 60 * minutes_in_day;
+    long seconds_in_day = 60 * minutes_in_day;
 
     printf("%d\n", days_in_month);
     printf("%d\n", days_in_year);
     printf("%d\n", minutes_in_day);
-    printf("%d\n", seconds_in_day);
+    printf("%ld\n", seconds_in_day);
 
     return 0;
 }
